Fix deleteFolderRecursive not deleting files

+++ src/background/action-creators/clear-filesystem.js
@@ -1,61 +1,43 @@
 import { fs } from '../index'
 
-async function deleteFolderRecursive (path) { // clears nonempty folders by recursion
+export async function deleteFolderRecursive (path) { // clears nonempty folders by recursion
   console.log(`deleteFolderRecursive called on ${path}`)
-  var contents = []
-  contents = await fs.promises.readdir(path)
+  let contents = await fs.promises.readdir(path)
   console.log(`deleteFolderRecursive: contents of ${path} are ${contents}`)
-  const handleIndividual = async function (file) {
-    const curPath = path + '/' + file
+  const handleIndividual = async function (item) {
+    const curPath = path + item
     const curPathStat = await fs.promises.lstat(curPath)
+    console.log(`handleIndividual called on ${curPath}`)
     if (curPathStat.isDirectory()) { // recurse
       // console.log(`deleteFolderRecursive: ${curPath} is a directory, recursing`)
-      await deleteFolderRecursive(curPath)
+      await deleteFolderRecursive(curPath + '/')
     } else { // delete file
-      await fs.promises.unlink(curPath).then(() => {
+      return fs.promises.unlink(curPath).then(async function () {
         console.log(`deleteFolderRecursive: unlink on ${curPath} succeeded `)
+        const contentsTesting = await fs.promises.readdir(path)
+        console.log(`AFTER UNLINKING deleteFolderRecursive: contents of ${path} are ${contentsTesting}`)
+        return true
       }, (err) => {
         console.log(`deleteFolderRecursive: unlink on ${curPath} failed with error ${err}`)
         throw err
       })
     }
   }
-  const contentPromises = []
-  for (let i = 0; i < contents.length; i++) {
-    contentPromises.push(handleIndividual(contents[i]))
+  const contentLength = contents.length
+  for (let i = 0; i < contentLength; i++) {
+    await handleIndividual(contents[i])
   }
-  await Promise.all(contentPromises) // all contents deleted, can now delete folder
-  await fs.promises.rmdir(path).then(() => {
+  console.log('for loop over')
+  if (path === '/') {
+    return true
+  }
+  return fs.promises.rmdir(path).then(() => {
     console.log(`deleteFolderRecursive: rmdir on ${path} succeeded`)
+    return true
   }, async function (err) {
     console.log(`deleteFolderRecursive: rmdir on ${path} failed with error ${err}`)
     const contents = await fs.promises.readdir(path)
     console.log(`current contents of of ${path} are ${contents}`)
     throw err
   })
-  console.log(`deleteFolderRecursive ending on ${path}`)
-}
-
-export async function clearFilesystem () {
-  const path = '/'
-  var contents = []
-  contents = await fs.promises.readdir(path).catch((err) => {
-    console.log(`clearFilesystem: readdir of ${path} failed with error ${err}`)
-    throw err
-  })
-  console.log(`clearFilesystem: contents of ${path} are ${contents}`)
-  contents.forEach(async function (file, index) {
-    const curPath = path + file
-    const curPathStat = await fs.promises.lstat(curPath)
-    if (curPathStat.isDirectory()) { // call delFoldRec
-      console.log(`clearFilesystem: ${curPath} is a directory, calling deleteFolderRecursive`)
-      await deleteFolderRecursive(curPath)
-    } else { // delete file
-      await fs.promises.unlink(curPath).catch((err) => {
-        console.log(`clearFilesystem: unlink on ${curPath} failed with error ${err}`)
-        throw err
-      })
-    }
-  })
-  return true
 }
